+++ lib/src/core/base/widget/appbar/leading_widget.dart
@@ -60,7 +60,6 @@ class LeadingWidget extends StatelessWidget {
               padding: EdgeInsets.zero,
               foregroundColor: iconColor,
               backgroundColor: iconBackgroundColor,
-
             ),
             onPressed: () => route?.canPop() ?? false ? route!.pop() : null,
           ),